dev-menu/requirements.txt
rumps>=0.4.0
pyobjc-framework-Cocoa>=10.0
pyinstaller>=6.0
